+++ extensions/hue/src/controlGroups.tsx
@@ -11,7 +11,7 @@ import useInputRateLimiter from "./hooks/useInputRateLimiter";
 import UnlinkAction from "./components/UnlinkAction";
 import SetScene from "./setScene";
 import { getColorFromLight, getLightsFromGroup } from "./helpers/hueResources";
-import { getTransitionTimeInMs, optimisticUpdates } from "./helpers/raycast";
+import { getTransitionTimeInMs, optimisticGroupUpdate } from "./helpers/raycast";
 import { calculateAdjustedBrightness, getClosestBrightness } from "./helpers/colors";
 import chroma from "chroma-js";
 import Style = Toast.Style;
@@ -36,7 +36,7 @@ export default function ControlGroups() {
         const groupColors = groupLights
           .filter((light) => uniqueColors.has(getColorFromLight(light)))
           .map((light) => getColorFromLight(light))
-          .sort((a, b) => chroma.hex(b).get("hsl.h") - chroma.hex(a).get("hsl.h"));
+          .sort((a, b) => chroma(b).get("hsl.h") - chroma(a).get("hsl.h"));
         return [group.id, groupColors];
       }),
     );
@@ -218,7 +218,10 @@ function SetBrightnessAction(props: { group: Group; groupedLight: GroupedLight; 
       title="Set Brightness"
       // This should be 0-100, but the API returns 0-254
       icon={getProgressIcon((props.groupedLight.dimming?.brightness ?? 0) / 254, Color.PrimaryText)}
-      shortcut={{ modifiers: ["cmd", "shift"], key: "b" }}
+      shortcut={{
+        macOS: { modifiers: ["cmd", "shift"], key: "b" },
+        Windows: { modifiers: ["ctrl", "shift"], key: "b" },
+      }}
     >
       {BRIGHTNESSES.map((brightness) => (
         <Action key={brightness} title={`${brightness}% Brightness`} onAction={() => props.onSet(brightness)} />
@@ -238,7 +241,10 @@ function IncreaseBrightnessAction(props: { group: Group; groupedLight: GroupedLi
   return (
     <Action
       title="Increase Brightness"
-      shortcut={{ modifiers: ["cmd", "shift"], key: "arrowUp" }}
+      shortcut={{
+        macOS: { modifiers: ["cmd", "shift"], key: "arrowUp" },
+        Windows: { modifiers: ["ctrl", "shift"], key: "arrowUp" },
+      }}
       icon={Icon.Plus}
       onAction={props.onIncrease}
     />
@@ -256,7 +262,10 @@ function DecreaseBrightnessAction(props: { group: Group; groupedLight: GroupedLi
   return (
     <Action
       title="Decrease Brightness"
-      shortcut={{ modifiers: ["cmd", "shift"], key: "arrowDown" }}
+      shortcut={{
+        macOS: { modifiers: ["cmd", "shift"], key: "arrowDown" },
+        Windows: { modifiers: ["ctrl", "shift"], key: "arrowDown" },
+      }}
       icon={Icon.Minus}
       onAction={props.onDecrease}
     />
@@ -264,7 +273,7 @@ function DecreaseBrightnessAction(props: { group: Group; groupedLight: GroupedLi
 }
 
 async function handleToggle(
-  { hueBridgeState, groupedLights, setGroupedLights, zones }: ReturnType<typeof useHue>,
+  { hueBridgeState, groupedLights, setGroupedLights, setLights, zones }: ReturnType<typeof useHue>,
   rateLimiter: ReturnType<typeof useInputRateLimiter>,
   groupLights: Light[],
   groupedLight: GroupedLight | undefined,
@@ -277,7 +286,7 @@ async function handleToggle(
     if (groupedLight === undefined) throw new Error("Light group not found.");
     if (!rateLimiter.canExecute()) return;
 
-    const changes = {
+    const changesToGroup = {
       on: { on: !groupedLight.on?.on },
       // No dynamics when toggling groups, as that causes the brightness to the set to the lowest possible level,
       //   even when only applied to toggling off.
@@ -299,34 +308,39 @@ async function handleToggle(
         )
         // Filter out undefined grouped lights
         .filter((zoneGroupedLight): zoneGroupedLight is GroupedLight => zoneGroupedLight !== undefined)
-        .map((zoneGroupedLight) => [zoneGroupedLight.id, changes]),
+        .map((zoneGroupedLight) => [zoneGroupedLight.id, changesToGroup]),
     )
       // Add the grouped light that triggered the action
-      .set(groupedLight.id, changes);
+      .set(groupedLight.id, changesToGroup);
 
-    const undoOptimisticGroupedLightsUpdate = optimisticUpdates(changesToGroupedLights, setGroupedLights);
-    await hueBridgeState.context.hueClient.updateGroupedLight(groupedLight, changes).catch((error: Error) => {
+    const changesToLights = new Map(groupLights.map((light) => [light.id, changesToGroup]));
+
+    const undoOptimisticUpdate = optimisticGroupUpdate(
+      changesToLights,
+      changesToGroupedLights,
+      setLights,
+      setGroupedLights,
+    );
+    await hueBridgeState.context.hueClient.updateGroupedLight(groupedLight, changesToGroup).catch((error: Error) => {
       if (error.message === 'device (grouped_light) is "soft off", command (.on) may not have effect') return;
-      undoOptimisticGroupedLightsUpdate();
+      undoOptimisticUpdate();
       throw error;
     });
 
     toast.style = Style.Success;
-    toast.title = groupedLight.on?.on ? `Turned ${group.metadata.name} off` : `Turned ${group.metadata.name} on`;
+    toast.title = `Turned ${groupedLight?.on?.on ? "off" : "on"} ${group.metadata.name}`;
     await toast.show();
   } catch (error) {
     console.error(error);
     toast.style = Style.Failure;
-    toast.title = groupedLight?.on?.on
-      ? `Failed turning ${group.metadata.name} off`
-      : `Failed turning ${group.metadata.name} on`;
+    toast.title = `Failed turning ${groupedLight?.on?.on ? "off" : "on"} ${group.metadata.name}`;
     toast.message = error instanceof Error ? error.message : undefined;
     await toast.show();
   }
 }
 
 async function handleSetBrightness(
-  { hueBridgeState, setLights }: ReturnType<typeof useHue>,
+  { hueBridgeState, setLights, setGroupedLights }: ReturnType<typeof useHue>,
   rateLimiter: ReturnType<typeof useInputRateLimiter>,
   groupLights: Light[],
   groupedLight: GroupedLight | undefined,
@@ -346,17 +360,22 @@ async function handleSetBrightness(
       dynamics: { duration: getTransitionTimeInMs() },
     };
 
+    const changesToGroupedLights = new Map([[groupedLight.id, changes]]);
     const changesToLights = new Map(groupLights.map((light) => [light.id, changes]));
-    const undoOptimisticUpdates = optimisticUpdates(changesToLights, setLights);
+
+    const undoOptimisticUpdate = optimisticGroupUpdate(
+      changesToLights,
+      changesToGroupedLights,
+      setLights,
+      setGroupedLights,
+    );
     await hueBridgeState.context.hueClient.updateGroupedLight(groupedLight, changes).catch((e: Error) => {
-      undoOptimisticUpdates();
+      undoOptimisticUpdate();
       throw e;
     });
 
     toast.style = Style.Success;
-    toast.title = `Set brightness of ${group.metadata.name} to ${(brightness / 100).toLocaleString("en", {
-      style: "percent",
-    })}.`;
+    toast.title = `Set brightness of ${group.metadata.name} to ${(brightness / 100).toLocaleString("en", { style: "percent" })}.`;
     await toast.show();
   } catch (error) {
     toast.style = Style.Failure;
@@ -367,7 +386,7 @@ async function handleSetBrightness(
 }
 
 async function handleBrightnessChange(
-  { hueBridgeState, lights, setLights }: ReturnType<typeof useHue>,
+  { hueBridgeState, lights, setLights, setGroupedLights }: ReturnType<typeof useHue>,
   rateLimiter: ReturnType<typeof useInputRateLimiter>,
   groupedLight: GroupedLight | undefined,
   group: Group,
@@ -390,25 +409,26 @@ async function handleBrightnessChange(
       dynamics: { duration: getTransitionTimeInMs() },
     };
 
+    const changesToGroupedLights = new Map([[groupedLight.id, changes]]);
     const changesToLights = new Map(getLightsFromGroup(lights, group).map((light) => [light.id, changes]));
-    const undoOptimisticUpdates = optimisticUpdates(changesToLights, setLights);
+
+    const undoOptimisticUpdate = optimisticGroupUpdate(
+      changesToLights,
+      changesToGroupedLights,
+      setLights,
+      setGroupedLights,
+    );
     await hueBridgeState.context.hueClient.updateGroupedLight(groupedLight, changes).catch((e: Error) => {
-      undoOptimisticUpdates();
+      undoOptimisticUpdate();
       throw e;
     });
 
     toast.style = Style.Success;
-    toast.title = `${direction === "increase" ? "Increased" : "Decreased"} brightness of ${group.metadata.name} to ${(
-      adjustedBrightness / 100
-    ).toLocaleString("en", {
-      style: "percent",
-    })}`;
+    toast.title = `${direction === "increase" ? "Increased" : "Decreased"} brightness of ${group.metadata.name} to ${(adjustedBrightness / 100).toLocaleString("en", { style: "percent" })}`;
     await toast.show();
   } catch (error) {
     toast.style = Style.Failure;
-    toast.title = `Failed ${direction === "increase" ? "increasing" : "decreasing"} brightness of ${
-      group.metadata.name
-    }`;
+    toast.title = `Failed ${direction === "increase" ? "increasing" : "decreasing"} brightness of ${group.metadata.name}`;
     toast.message = error instanceof Error ? error.message : undefined;
     await toast.show();
   }