+++ extensions/obsidian/src/components/NoteList/NoteListItem/NoteListItem.tsx
@@ -15,6 +15,38 @@ import { useNoteContent } from "../../../utils/hooks";
 import { useState } from "react";
 import { Note, ObsidianVault, ObsidianUtils } from "@/obsidian";
 
+interface NoteListItemMetadataProps {
+  content: string;
+  note: Note;
+  vault: ObsidianVault;
+}
+
+function NoteListItemMetadata({ content, note, vault }: NoteListItemMetadataProps) {
+  const tags = ObsidianUtils.getAllTags(content);
+  const url = ObsidianUtils.getProperty(content, "url");
+
+  return (
+    <List.Item.Detail.Metadata>
+      <List.Item.Detail.Metadata.Label title="Character Count" text={content.length.toString()} />
+      <List.Item.Detail.Metadata.Label title="Word Count" text={wordCount(content).toString()} />
+      <List.Item.Detail.Metadata.Label title="Reading Time" text={readingTime(content).toString() + " min read"} />
+      <List.Item.Detail.Metadata.TagList title="Tags">
+        {tags.map((tag) => (
+          <List.Item.Detail.Metadata.TagList.Item key={tag} text={tag} />
+        ))}
+      </List.Item.Detail.Metadata.TagList>
+      {url && <List.Item.Detail.Metadata.Link target={url} text="View" title="URL" />}
+      <List.Item.Detail.Metadata.Separator />
+      <List.Item.Detail.Metadata.Label title="Creation Date" text={createdDateFor(note).toLocaleDateString()} />
+      <List.Item.Detail.Metadata.Label title="File Size" text={fileSizeFor(note).toFixed(2) + " KB"} />
+      <List.Item.Detail.Metadata.Label
+        title="Note Path"
+        text={trimPathToMaxLength(note.path.split(vault.path)[1], 55)}
+      />
+    </List.Item.Detail.Metadata>
+  );
+}
+
 export function NoteListItem(props: {
   note: Note;
   vault: ObsidianVault;
@@ -38,61 +70,9 @@ export function NoteListItem(props: {
   // Create a modified note object with the current bookmark state
   const updatedNote = { ...note, bookmarked: isBookmarked };
 
-  function TagList() {
-    return null;
-    // if (note.tags.length > 0) {
-    //   return (
-    //     <List.Item.Detail.Metadata.TagList title="Tags">
-    //       {note.tags.map((tag) => (
-    //         <List.Item.Detail.Metadata.TagList.Item key={tag} text={tag} />
-    //       ))}
-    //     </List.Item.Detail.Metadata.TagList>
-    //   );
-    // } else {
-    //   return null;
-    // }
-  }
+  if (noteHasBeenMoved) return null;
 
-  function Link() {
-    if (!noteContent) return null;
-    const url = ObsidianUtils.getProperty(noteContent, "url");
-    if (url) {
-      return <List.Item.Detail.Metadata.Link target={url} text="View" title="URL" />;
-    } else {
-      return null;
-    }
-  }
-
-  function renderMetadata() {
-    if (!noteContent || !pref.showMetadata) {
-      return <></>;
-    }
-
-    return (
-      <List.Item.Detail.Metadata>
-        <List.Item.Detail.Metadata.Label title="Character Count" text={noteContent.length.toString()} />
-        <List.Item.Detail.Metadata.Label title="Word Count" text={wordCount(noteContent).toString()} />
-        <List.Item.Detail.Metadata.Label
-          title="Reading Time"
-          text={readingTime(noteContent).toString() + " min read"}
-        />
-        <TagList />
-        <Link />
-        <List.Item.Detail.Metadata.Separator />
-        <List.Item.Detail.Metadata.Label
-          title="Creation Date"
-          text={createdDateFor(updatedNote).toLocaleDateString()}
-        />
-        <List.Item.Detail.Metadata.Label title="File Size" text={fileSizeFor(updatedNote).toFixed(2) + " KB"} />
-        <List.Item.Detail.Metadata.Label
-          title="Note Path"
-          text={trimPathToMaxLength(updatedNote.path.split(vault.path)[1], 55)}
-        />
-      </List.Item.Detail.Metadata>
-    );
-  }
-
-  return !noteHasBeenMoved ? (
+  return (
     <List.Item
       title={updatedNote.title}
       id={updatedNote.path}
@@ -106,33 +86,41 @@ export function NoteListItem(props: {
         },
       ]}
       detail={
-        <List.Item.Detail
-          isLoading={isLoading}
-          markdown={filterContent(noteContent ?? "")}
-          metadata={noteContent ? renderMetadata() : null}
-        />
+        noteContent && (
+          <List.Item.Detail
+            isLoading={isLoading}
+            markdown={ObsidianUtils.renderCallouts(filterContent(noteContent))}
+            metadata={
+              noteContent && pref.showMetadata ? (
+                <NoteListItemMetadata note={note} content={noteContent} vault={vault} />
+              ) : null
+            }
+          />
+        )
       }
       actions={
-        <ActionPanel>
-          <OpenNoteActions note={{ content: noteContent ?? "", ...updatedNote }} vault={vault} />
-          <NoteActions
-            note={{ content: noteContent ?? "", ...updatedNote }}
-            vault={vault}
-            onNoteAction={(actionType) => {
-              switch (actionType) {
-                case "bookmark":
-                  setIsBookmarked(true);
-                  break;
-                case "unbookmark":
-                  setIsBookmarked(false);
-                  break;
-              }
-            }}
-            onNoteUpdated={onNoteUpdated}
-            onDelete={onDelete}
-          />
-        </ActionPanel>
+        noteContent && (
+          <ActionPanel>
+            <OpenNoteActions note={{ content: noteContent, ...updatedNote }} vault={vault} />
+            <NoteActions
+              note={{ content: noteContent, ...updatedNote }}
+              vault={vault}
+              onNoteAction={(actionType) => {
+                switch (actionType) {
+                  case "bookmark":
+                    setIsBookmarked(true);
+                    break;
+                  case "unbookmark":
+                    setIsBookmarked(false);
+                    break;
+                }
+              }}
+              onNoteUpdated={onNoteUpdated}
+              onDelete={onDelete}
+            />
+          </ActionPanel>
+        )
       }
     />
-  ) : null;
+  );
 }